+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### v1.1.0
+* Make mod localizable
+
 ### v1.0.3
 * Fix: correctly position relic popup if there are multiple pages of relics.
 * Fix: relic sell price tip no longer moves off-screen for relics on the right of the screen.